+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * update `freemius/wordpress-sdk`
+* update `afragen/wp-dismiss-notice`
 
 #### 12.1.3 / 2023-03-20
 * improved setting/default of `$options['bypass_background_processing']`