+++ CHANGES.md
@@ -1,6 +1,7 @@
 #### [unreleased]
 * update `freemius/wordpress-sdk`
 * update `afragen/wp-dismiss-notice`
+* don't save unused data from `API_Common::parse_release_asset()`
 
 #### 12.1.3 / 2023-03-20
 * improved setting/default of `$options['bypass_background_processing']`