+++ CHANGES.md
@@ -2,6 +2,7 @@
 * update `freemius/wordpress-sdk`
 * update `afragen/wp-dismiss-notice`
 * don't save unused data from `API_Common::parse_release_asset()`
+* update Freemius uninstall to use previous `uninstall.php` now renamed
 
 #### 12.1.3 / 2023-03-20
 * improved setting/default of `$options['bypass_background_processing']`