+++ CHANGES.md
@@ -7,6 +7,7 @@
 * composer update
 * update `REST_API::get_plugins_api_data()` to return response without download link using boolean value in `download` query arg
 * hide Freemius menus with `gu_hide_settings` filter
+* more specific hiding of Git Updater settings
 
 #### 12.1.3 / 2023-03-20
 * improved setting/default of `$options['bypass_background_processing']`